Add pygame2exe.py, fix build & credits, reduce some obvious bloat, add icon.
Add pygame2exe.py for Windows builds, based on sample from pygame.org
Fix py2exe build "error: could not create 'build': Cannot create a file when that file already exists"
Fix Credits: "TypeError: integer argument expected, got float"
Remove .xvpics (unnecessary thumbnails)
While I'm at it, I've recompressed the pngs in the "clean" theme.
Add appicon.ico (from PortableApps.com release)

+++ animator.py
@@ -80,7 +80,7 @@ class AnimatedTurnObject:
 	def getSurface(self, time):
 		offset = (time-self.delay)
 
-		return pygame.transform.scale(self.image, (math.sin(self.delta*offset)*self.image.get_width(), self.image.get_height()))
+		return pygame.transform.scale(self.image, (int(math.sin(self.delta*offset)*self.image.get_width()), self.image.get_height()))
 
 
 	def visable(self, time):
